requirements.txt
absl-py==0.15.0
alembic==1.7.6
argon2-cffi==21.1.0
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
clang==5.0
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
cryptography==35.0.0
cycler==0.11.0
databricks-cli==0.16.4
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
dlx==1.0.4
docker==5.0.3
docplex==2.22.213
entrypoints==0.3
fastdtw==0.3.4
fastjsonschema==2.15.1
flake8==4.0.1
Flask==2.0.3
flatbuffers==1.12
gast==0.3.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.32.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==1.1.1
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.2.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras==2.3.1
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lxml==4.6.4
Mako==1.1.6
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mlflow==1.24.0
more-itertools==8.10.0
mpmath==1.2.1
multitasking==0.0.9
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
ntlm-auth==1.5.0
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.6.0
pluggy==1.0.0
ply==3.11
prometheus-client==0.12.0
prometheus-flask-exporter==0.19.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.8.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
Pygments==2.10.0
pylatexenc==2.10
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==7.1.1
python-constraint==1.4.0
python-dateutil==2.8.2
pytz==2021.3
pywin32==227
pywinpty==1.1.5
PyYAML==6.0
pyzmq==22.3.0
qiskit==0.31.0
qiskit-aer==0.9.1
qiskit-aqua==0.9.5
qiskit-ibmq-provider==0.17.0
qiskit-ignis==0.6.0
qiskit-terra==0.18.3
qtconsole==5.1.1
QtPy==1.11.2
Quandl==3.6.2
querystring-parser==1.2.4
requests==2.26.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
retworkx==0.10.2
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.1
Send2Trash==1.8.0
six==1.15.0
smmap==5.0.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
sympy==1.9
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tomli==2.0.1
torch==1.10.2
torch-summary==1.4.5
torchaudio==0.10.2
torchsummary==1.5.1
torchvision==0.11.3
tornado==6.1
traitlets==5.1.1
tweedledum==1.1.1
typing-extensions==3.7.4.3
urllib3==1.26.7
waitress==2.1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wincertstore==0.2
wrapt==1.12.1
yfinance==0.1.64
zipp==3.6.0
